+++ core/src/reference/mod.rs
@@ -4,8 +4,10 @@ mod command;
 
 use initiative_macros::reference_enum;
 
-reference_enum!(Spell);
-
 reference_enum!(Item);
 
 reference_enum!(ItemCategory);
+
+reference_enum!(Spell);
+
+reference_enum!(MagicItem);